Feat: 미니 지도 파일 분리 시도
